# Remove unused code in parsing and speed parsing up