The `NanoDate(::AbstractString, ::DateFormat)` constructor previously
created a few variables which are never used, so remove these lines.
These lines are actually responsible for about half the running time of
calls to the constructor, so this change doubles the speed of parsing.

Also remove the unused `separate_offset` methods.