@anowacki 2024-08-15
Before (`v1.0.2`):
```julia-repl
julia> @benchmark NanoDate("1234-01-02T01:23:45.123456789+01:23")
BenchmarkTools.Trial: 10000 samples with 3 evaluations.
 Range (min … max):  8.617 μs … 28.051 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     9.091 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.873 μs ±  1.543 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁▆██▇▆▄▂      ▁▃▃▃▃▂▁ ▁▁▁▂▂▃▃▃▂▂▁▁▁▁▁▂▁▁▁                  ▂
  █████████▇▇▇█████████████████████████████████▆▇█▇▇▇▇▇▇▆▆▅▅ █
  8.62 μs      Histogram: log(frequency) by time     14.9 μs <

 Memory estimate: 2.66 KiB, allocs estimate: 52.
```

After (`016ccae`):
```julia-repl
julia> @benchmark NanoDate("1234-01-02T01:23:45.123456789+01:23")
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  5.056 μs …  10.450 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.241 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.297 μs ± 325.704 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▃██▅                                                       
  ▃▆█████▅▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▁▁▁▂▂▂▁▂▂▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▂▂ ▃
  5.06 μs         Histogram: frequency by time        7.51 μs <

 Memory estimate: 1.88 KiB, allocs estimate: 43.
```